Remove support for deserializing MapData with map key other than "step" (#4719)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4719

D81363344 / Ax #4255 disallowed having map keys other than "step", providing deserialization support, with warnings, when a map key other than "step" is deserialized. This was released in Ax 1.1.2; we are now on Ax 1.2.1. Since there have been warnings for a while about this, it should be save to remove the deserialization support.

With this change, if a progression was saved with some other column name, such as "epoch," it will be deserialized as another MapData column, with NaN imputed for "step." When this logic goes away in D89820078, it will be deserialized as Data with no "step" column.

Reviewed By: saitcakmak

Differential Revision: D89888874

fbshipit-source-id: be3de0081a2f12f725961c1d36e65e142bfca511